refactor: Extract `initHtmlContent` from `CozyProxyWebView.js`
